rename modules for green tabs
